# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Medical Appointment API
  version: 1.0.0
  description: API para agendamiento de citas médicas

servers:
  - url: https://0a7pva8mxc.execute-api.us-east-1.amazonaws.com
    description: AWS API Gateway

paths:
  /appointments:
    post:
      summary: Crear una nueva cita médica
      description: Endpoint para crear una nueva cita médica
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        202:
          description: Cita aceptada para procesamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAppointmentResponse"
        400:
          description: Datos de entrada inválidos
        500:
          description: Error interno del servidor

    get:
      summary: Obtener todas las citas médicas
      description: Endpoint para listar todas las citas médicas
      operationId: getAllAppointments
      responses:
        200:
          description: Lista de citas médicas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        500:
          description: Error interno del servidor

  /appointments/{insuredId}:
    get:
      summary: Obtener citas por ID de asegurado
      description: Endpoint para obtener citas médicas de un asegurado específico
      operationId: getAppointmentsByInsuredId
      parameters:
        - name: insuredId
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{5}$'
          example: "12346"
      responses:
        200:
          description: Lista de citas del asegurado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        404:
          description: No se encontraron citas para este asegurado
        500:
          description: Error interno del servidor

components:
  schemas:
    CreateAppointmentRequest:
      type: object
      required:
        - insuredId
        - scheduleId
        - countryISO
      properties:
        insuredId:
          type: string
          pattern: '^\d{5}$' # Asegura que el ID sea un número de 5 dígitos
          example: "12349"
          description: ID del asegurado (5 dígitos)
        scheduleId:
          type: integer
          example: 1
          description: ID del horario disponible
        countryISO:
          type: string
          enum: [PE, CL] # Los países permitidos son PE o CL
          example: "CL"
          description: País del asegurado (PE-Perú, CL-Chile)
        status:
          type: string
          enum: [pending, completed]
          example: "pending"
          description: Estado de la cita
        createdAt:
          type: string
          format: date-time
          example: "2025-04-20T10:00:00Z"
          description: Fecha de creación de la cita
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-20T10:00:00Z"
          description: Fecha de última actualización de la cita

    CreateAppointmentResponse:
      type: object
      properties:
        message:
          type: string
          example: "Appointment scheduling in process"
        appointmentId:
          type: string
          example: "appt-vbobp755dx"

    Appointment:
      type: object
      properties:
        appointmentId:
          type: string
          example: "appt-pk4jwp48cb"
        insuredId:
          type: string
          example: "12346"
        scheduleId:
          type: integer
          example: 1
        countryISO:
          type: string
          example: "PE"
        status:
          type: string
          enum: [pending, completed]
          example: "completed"
        createdAt:
          type: string
          format: date-time
          example: "2025-04-28T01:32:57.644Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-28T01:32:57.784Z"